[screen01] Lesson 1 Screen01 - Exploring Graphics through Framebuffer
Add two libraries for working with the GPU to display graphics. The first,
mailbox.s, helps with reading and writing messages between CPU and GPU. The
second, framebuffer.s, helps construct a framebuffer message to pass to the GPU.

The main program causes a colorful gradient to be displayed on screen.